requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
asgiref=3.5.2=pypi_0
brotlipy=0.7.0=pypi_0
ca-certificates=2022.10.11=haa95532_0
certifi=2020.12.5=pypi_0
cffi=1.15.1=pypi_0
charset-normalizer=2.1.1=pypi_0
colorama=0.4.6=pypi_0
django=3.2=pypi_0
django-appconf=1.0.4=pypi_0
django-ckeditor=6.4.0=pypi_0
django-imagekit=4.0.2=pypi_0
django-js-asset=2.0.0=pypi_0
django-redis=4.12.1=pypi_0
gevent=21.1.2=pypi_0
greenlet=1.0.0=pypi_0
grequests=0.6.0=pypi_0
idna=3.4=pypi_0
jieba=0.42.1=pypi_0
jinja2=2.11.3=pypi_0
libffi=3.4.2=hd77b12b_6
markupsafe=1.1.1=pypi_0
openssl=1.1.1s=h2bbff1b_0
pilkit=2.0=pypi_0
pillow=8.1.2=pypi_0
pip=22.2.2=py38haa95532_0
prettytable=2.1.0=pypi_0
pycparser=2.21=pypi_0
pyecharts=1.9.0=pypi_0
pymysql=1.0.2=pypi_0
python=3.8.15=h6244533_2
pytz=2022.6=pypi_0
qrcode=7.3.1=pypi_0
redis=3.5.3=pypi_0
requests=2.28.1=pypi_0
setuptools=65.5.0=py38haa95532_0
simplejson=3.17.2=pypi_0
six=1.16.0=pypi_0
some-package=0.1=pypi_0
sqlite=3.40.0=h2bbff1b_0
sqlparse=0.4.3=pypi_0
urllib3=1.26.13=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
wincertstore=0.2=py38haa95532_2
zope-event=4.5.0=pypi_0
zope-interface=5.3.0=pypi_0
